move replaceGender to GenderHelper
