+++ tests/Unit/Helpers/GenderHelperTest.php
@@ -3,7 +3,10 @@
 namespace Tests\Unit\Helpers;
 
 use Tests\FeatureTestCase;
-use App\Helpers\GendersHelper;
+use App\Helpers\GenderHelper;
+use App\Models\Account\Account;
+use App\Models\Contact\Contact;
+use App\Models\Contact\Gender;
 
 class GenderHelperTest extends FeatureTestCase
 {
@@ -12,12 +15,40 @@ class GenderHelperTest extends FeatureTestCase
     {
         $this->signIn();
 
-        $genders = GendersHelper::getGendersInput();
+        $genders = GenderHelper::getGendersInput();
 
         $this->assertCount(4, $genders);
         $this->assertEquals([
             'id' => '',
             'name' => 'No gender',
         ], $genders[0]);
+    }
+
+    /** @test */
+    public function it_replaces_gender_with_another_gender()
+    {
+        $account = factory(Account::class)->create();
+        $male = factory(Gender::class)->create([
+            'account_id' => $account->id,
+        ]);
+        $female = factory(Gender::class)->create([
+            'account_id' => $account->id,
+        ]);
+
+        factory(Contact::class, 2)->create([
+            'account_id' => $account->id,
+            'gender_id' => $male
+        ]);
+        factory(Contact::class)->create([
+            'account_id' => $account->id,
+            'gender_id' => $female
+        ]);
+
+        $account->replaceGender($male, $female);
+
+        $this->assertEquals(
+            3,
+            $female->contacts->count()
+        );
     }
 }
